Extract loaders from ClassMetadataFactory
