+++ Resources/config/metadata.xml
@@ -10,17 +10,11 @@
 
     <services>
         <service id="dunglas_json_ld_api.mapping.class_metadata_factory" class="Dunglas\JsonLdApiBundle\Mapping\ClassMetadataFactory">
-            <argument type="service" id="dunglas_json_ld_api.resource_collection" />
-            <argument type="service" id="property_info" />
-            <argument type="service" id="validator" on-invalid="ignore" />
-            <argument type="service" id="serializer.mapping.class_metadata_factory" on-invalid="ignore" />
+            <argument type="service" id="dunglas_json_ld_api.mapping.loaders.chain" />
             <argument>null</argument>
         </service>
 
         <service id="dunglas_json_ld_api.cache_warmer.metadata" class="Dunglas\JsonLdApiBundle\CacheWarmer\MetadataCacheWarmer">
-            <argument type="service" id="dunglas_json_ld_api.resource_collection" />
-            <argument type="service" id="dunglas_json_ld_api.mapping.class_metadata_factory" />
-
             <tag name="kernel.cache_warmer" priority="0" />
         </service>
 
@@ -28,6 +22,34 @@
             <call method="setNamespace">
                 <argument>%dunglas_json_ld_api.mapping.cache.prefix%</argument>
             </call>
+        </service>
+
+        <!-- Loaders -->
+        <service id="dunglas_json_ld_api.mapping.loaders.chain" class="Dunglas\JsonLdApiBundle\Mapping\Loader\LoaderChain" public="false">
+            <argument type="collection">
+                <argument type="service" id="dunglas_json_ld_api.mapping.loaders.attributes" />
+                <argument type="service" id="dunglas_json_ld_api.mapping.loaders.validator_metadata" />
+                <argument type="service" id="dunglas_json_ld_api.mapping.loaders.phpdoc" />
+                <argument type="service" id="dunglas_json_ld_api.mapping.loaders.annotation" />
+            </argument>
+        </service>
+
+        <service id="dunglas_json_ld_api.mapping.loaders.attributes" class="Dunglas\JsonLdApiBundle\Mapping\Loader\AttributesLoader" public="false">
+            <argument type="service" id="property_info" />
+            <argument type="service" id="dunglas_json_ld_api.resource_collection" />
+            <argument type="service" id="serializer.mapping.class_metadata_factory" on-invalid="ignore" />
+        </service>
+
+        <service id="dunglas_json_ld_api.mapping.loaders.validator_metadata" class="Dunglas\JsonLdApiBundle\Mapping\Loader\ValidatorMetadataLoader" public="false">
+            <argument type="service" id="validator" />
+        </service>
+
+        <service id="dunglas_json_ld_api.mapping.loaders.phpdoc" class="Dunglas\JsonLdApiBundle\Mapping\Loader\PhpDocLoader" public="false">
+            <argument type="service" id="property_info" />
+        </service>
+
+        <service id="dunglas_json_ld_api.mapping.loaders.annotation" class="Dunglas\JsonLdApiBundle\Mapping\Loader\AnnotationLoader" public="false">
+            <argument type="service" id="annotations.reader" />
         </service>
     </services>
