+++ Tests/Behat/TestBundle/Entity/Dummy.php
@@ -13,6 +13,7 @@ namespace Dunglas\JsonLdApiBundle\Tests\Behat\TestBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
+use Dunglas\JsonLdApiBundle\Annotation\Iri;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
@@ -37,6 +38,7 @@ class Dummy
      *
      * @ORM\Column
      * @Assert\NotBlank
+     * @Iri("http://schema.org/name")
      */
     private $name;
     /**
