+++ Util/Reflection.php
@@ -1,0 +1,39 @@
+<?php
+
+/*
+ * This file is part of the DunglasJsonLdApiBundle package.
+ *
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Dunglas\JsonLdApiBundle\Util;
+
+/**
+ * Reflection utils.
+ *
+ * @author Kévin Dunglas <dunglas@gmail.com>
+ */
+trait Reflection
+{
+    /**
+     * Gets the {@see \ReflectionProperty} from the class or its parent.
+     *
+     * @param \ReflectionClass $reflectionClass
+     * @param string           $attributeName
+     *
+     * @return \ReflectionProperty
+     */
+    private function getReflectionProperty(\ReflectionClass $reflectionClass, $attributeName)
+    {
+        if ($reflectionClass->hasProperty($attributeName)) {
+            return $reflectionClass->getProperty($attributeName);
+        }
+
+        if ($parent = $reflectionClass->getParentClass()) {
+            return $this->getReflectionProperty($parent, $attributeName);
+        }
+    }
+}
